focus no field e abrir teclado em importar tema

+++ app/src/main/java/br/ufpb/dcx/appalpha/control/service/ThemeSqlService.java
@@ -137,7 +137,7 @@ public class ThemeSqlService {
         return themes;
     }
 
-    public boolean existsByApiId(Long id) {
+    public boolean existsByApiId(int id) {
         String selectQuery = "SELECT exists(SELECT id FROM " + DbHelper.THEMES_TABLE + " WHERE apiId = ? LIMIT 1)";
 
         Cursor cursor = readableDb.rawQuery(selectQuery, new String[]{Long.toString(id)});
